% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omia-package.R
\docType{package}
\name{omia}
\alias{omia-package}
\alias{omia}
\title{omia: Epidemiological and Statistical Workflows}
\description{
The \code{omia} package provides practical tools for incidence data preparation,
time-varying reproduction number estimation, growth model fitting,
model comparison, simulation, and reporting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OJWatson/omia}
}

}
\author{
\strong{Maintainer}: Kana Contributor \email{kana@example.com}

}
\keyword{internal}
