% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rt_estimates.R
\name{plot_rt_estimates}
\alias{plot_rt_estimates}
\title{Plot time-varying reproduction number estimates}
\usage{
plot_rt_estimates(
  rt_data,
  date_col = "date_end",
  mean_col = "mean_r",
  lower_col = "lower_95",
  upper_col = "upper_95"
)
}
\arguments{
\item{rt_data}{A tibble, typically from \code{\link[=estimate_rt_epiestim]{estimate_rt_epiestim()}}.}

\item{date_col}{Name of date column.}

\item{mean_col}{Name of mean Rt column.}

\item{lower_col}{Name of lower interval column.}

\item{upper_col}{Name of upper interval column.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plots mean \code{R_t} with 95\% uncertainty band.
}
